Pull Snapcraft v3.9.x (#16)
This PR is mostly to pull a newer version of Snapcraft. At the same
thime though, we're updating the base image for core18 to the CircleCI
November base. We're not switching to the January base yet as there's
more work and communication I need to do around not defaulting to the
root user anymore.